Increase MAX_HOTSTRING_LENGTH from 40 to 400
(I haven't tested this change, but hopefully it just works, given that this is just an arbitrary limit.)

(If the documentation, such as https://www.autohotkey.com/docs/v2/Hotstrings.htm#remarks, also needs to be changed, I also haven't done so.)

The comment for this constant (which I have kept for the new quantity ha ha) mentions "hard to imagine a need for more than this"; after years, we have found such a need to exceed 40. https://github.com/alexmick/emoji-to-ahk/blob/f7471fceda3f23cf36a83ba150c6a22fbb2796f1/custom_short_names.py#L2 . In this project that lets you type emoji based on their name (in a bbcode-style, :like_this:), some of the emoji names exceed 40 characters, and so something awkward has to be done to deal with that; why not allow the project to be perfect, instead? :)

I don't think those emoji names are necessarily the official unicode names for the characters, but I also want to be able to type such unicode names as bbcode-style hotstrings for another project of mine (which I may never get to). I don't think there's any limit on how long a unicode character name can be, but according to this random Quora answer from two years ago, the longest unicode character name is is 83 characters long https://www.quora.com/Which-are-the-longest-Unicode-characters/answer/Thomas-Phinney . 400 should be ample for this as well.

So why 400?

1. It's amusing simple to just add a 0.
2. It's still obviously an arbitrary (instead of technical) limit, unlike my favorite numbers, which are all powers of two.
3. An order of magnitude greater than our current need, so inshallah we will never have to increase it again.
4. Does not exceed INT_MAX